refactor: kill db/index
we will require importing from functionality-specific sub-files. we
refactor the call sites.
