+++ src/server/routes/users.js
@@ -3,7 +3,7 @@
 const express = require('express')
 const router = express.Router()
 const { requireAdminUser } = require('../access-helpers')
-const { createUser, user: getUser, updateUser } = require('../db')
+const { createUser, user: getUser, updateUser } = require('../db/users')
 const { agencyById } = require('../db/agencies')
 const { sendWelcomeEmail } = require('../lib/email')
 const _ = require('lodash-checkit')
